Support multiple openai versions in python (#27)
This change updates autoevals to support both the v0 and v1 versions of
OpenAI's python library. To accomplish this, we:

* Remove `openai` as a dependency. This enables autoevals to be included
flexibly as a dependency in systems which use either version of openai.
* Wrap the completion method and rate limit exception in a way that
works for either SDK version
* Switch to using the wrappers built into the `braintrust` sdk instead
of a custom one (except for cached requests...).

Tested using end-to-end scripts with both versions of python.