feature: configurable (OpenAI) client
All public interfaces that work with OpenAI transitively
now has an optional AutoEvalCient client option. If not provided,
we'll prepare and handle the OpenAI client as we do already.

At the moment you need to pass the client in each call, but
in a follow-up commit you'll be able to set it globally.