add test coverage for prepare_openai
Discovered that the client.openai should have been the openai instance.

Also, that the wrapping of the client should not be automatic as the
attached methods may not be the right ones. i.e. if the client.openai is
unwrapped, then we wrap in the prepare_openai the attached methods in
the client.complete would be the unwrapped ones vs. the new wrapped one.